# TPE Arquitectura Web Final/openapi.yaml
openapi: 3.0.1
info:
  title: API Monopatines Electricos
  description: Esta API se encarga del manego de un sistema de monopatines electricos
  version: 1.0.0
servers:
- url: http://localhost:8080
...
paths:
...
  /api/authenticate:
    post:
      tags:
      - Security
      summary: auntentifica un usuario
      responses: 
        200: 
          description: (ok) la peticion fue exitosa
        400: 
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
  /api/register:
    post:
      tags:
      - Security
      summary: registra un usuario
      responses: 
        200: 
          description: (ok) la peticion fue exitosa
        400: 
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/ServerError'
components: 
  responses: 
    Unauthorized:
      description: (Unauthorized) no tienes autorizacion 
    NotFound:
      description: (NotFound) elemento no encontrado
    BadRequest:
      description: (BadRequest) datos invalidos
    ServerError:
      description: (ServerError) error en el servidor
...